Atualizaçoes de tabela

+++ PizzariaManagementSystem/src/telas/DeletarUsuarios.java
@@ -5,7 +5,9 @@
 package telas;
 import java.sql.*;
 import conexaodatabase.ModuloConexao;
+import java.awt.HeadlessException;
 import javax.swing.JOptionPane;
+import javax.swing.table.DefaultTableModel;
 /**
  *
  * @author autologon
@@ -19,8 +21,25 @@ public class DeletarUsuarios extends javax.swing.JInternalFrame {
      */
     public DeletarUsuarios() {
         initComponents();
+        conexao = ModuloConexao.connector();
+        try {
+            String sql = "select * from usuarios ";
+            pst = conexao.prepareStatement(sql);
+            rs = pst.executeQuery();
+            while(rs.next()){
+                String[] nomes = {rs.getString(2)};
+                DefaultTableModel table = (DefaultTableModel)jTable1.getModel();
+                table.addRow(nomes);
+            }
+            conexao.close();
+          
+        } catch (SQLException e) {
+        }
+        
+        
+        
     }
-
+ 
     /**
      * This method is called from within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
@@ -33,6 +52,8 @@ public class DeletarUsuarios extends javax.swing.JInternalFrame {
         jLabel1 = new javax.swing.JLabel();
         delNome = new javax.swing.JTextField();
         jButton1 = new javax.swing.JButton();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        jTable1 = new javax.swing.JTable();
 
         setClosable(true);
         setIconifiable(true);
@@ -49,24 +70,51 @@ public class DeletarUsuarios extends javax.swing.JInternalFrame {
             }
         });
 
+        jTable1.setModel(new javax.swing.table.DefaultTableModel(
+            new Object [][] {
+
+            },
+            new String [] {
+                "Nome"
+            }
+        ) {
+            Class[] types = new Class [] {
+                java.lang.String.class
+            };
+
+            public Class getColumnClass(int columnIndex) {
+                return types [columnIndex];
+            }
+        });
+        jTable1.setEnabled(false);
+        jTable1.setOpaque(false);
+        jScrollPane1.setViewportView(jTable1);
+
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
         layout.setHorizontalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(69, 69, 69)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, 96, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addGroup(layout.createSequentialGroup()
-                        .addComponent(jLabel1)
-                        .addGap(29, 29, 29)
-                        .addComponent(delNome, javax.swing.GroupLayout.PREFERRED_SIZE, 276, javax.swing.GroupLayout.PREFERRED_SIZE)))
-                .addContainerGap(135, Short.MAX_VALUE))
+                        .addGap(69, 69, 69)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(jButton1, javax.swing.GroupLayout.PREFERRED_SIZE, 96, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addGroup(layout.createSequentialGroup()
+                                .addComponent(jLabel1)
+                                .addGap(29, 29, 29)
+                                .addComponent(delNome, javax.swing.GroupLayout.PREFERRED_SIZE, 276, javax.swing.GroupLayout.PREFERRED_SIZE))))
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(46, 46, 46)
+                        .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addContainerGap(77, Short.MAX_VALUE))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addGap(165, 165, 165)
+                .addGap(16, 16, 16)
+                .addComponent(jScrollPane1, javax.swing.GroupLayout.PREFERRED_SIZE, 113, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(36, 36, 36)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(delNome, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE))
@@ -81,12 +129,23 @@ public class DeletarUsuarios extends javax.swing.JInternalFrame {
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
        conexao = ModuloConexao.connector();
         try {
-            String sql = "delete from usuarios where nome = ?";
+            String sql = "select * from usuarios where Nome=?";
             pst = conexao.prepareStatement(sql);
-            pst.setString(1, delNome.getText());
-            pst.executeUpdate();
-            JOptionPane.showMessageDialog(null, "usuario deletado com sucesso.");
-        } catch (Exception e) {
+            pst.setString(1,delNome.getText());
+            rs = pst.executeQuery();
+            if (rs.next()) {
+                String sql1 = "delete from usuarios where nome = ?";
+                pst = conexao.prepareStatement(sql1);
+                pst.setString(1, delNome.getText());
+                pst.executeUpdate();
+                JOptionPane.showMessageDialog(null, "Usuário deletado com sucesso.");
+
+            }
+            else{
+                JOptionPane.showMessageDialog(null, "Usuário já não existe");
+            }
+            
+        } catch (HeadlessException | SQLException e) {
             JOptionPane.showMessageDialog(null, e);
         }
     }//GEN-LAST:event_jButton1ActionPerformed
@@ -96,5 +155,7 @@ public class DeletarUsuarios extends javax.swing.JInternalFrame {
     private javax.swing.JTextField delNome;
     private javax.swing.JButton jButton1;
     private javax.swing.JLabel jLabel1;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JTable jTable1;
     // End of variables declaration//GEN-END:variables
 }
